+++ src/main/java/com/Harbinger/Spore/Client/Models/InfectedEvokerModel.java
@@ -140,7 +140,7 @@ public class InfectedEvokerModel<T extends InfectedEvoker> extends EntityModel<T
 	@Override
 	public void setupAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
 		if (entity.attackAnim > 0) {
-			float f1 = 1.0F - (float) Mth.abs(10 - 2 * entity.attackAnim) / 10.0F;
+			float f1 = 1.0F - Mth.abs(10 - 2 * entity.attackAnim) / 10.0F;
 			this.RightArm.xRot = -140f;
 			this.LeftArm.xRot = -140f;
 			this.RightArm.yRot = Mth.sin(ageInTicks/6)/8 + Mth.sin(f1) * 2.0F;
