+++ src/main/java/com/Harbinger/Spore/Client/Models/SlasherModel.java
@@ -123,7 +123,7 @@ public class SlasherModel<T extends Slasher> extends EntityModel<T> {
 	public void setupAnim(T entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
 
 		if (entity.attackAnim > 0) {
-			float f1 = 1.0F - (float) Mth.abs(10 - 2 * entity.attackAnim) / 10.0F;
+			float f1 = 1.0F - Mth.abs(10 - 2 * entity.attackAnim) / 10.0F;
 			this.Marm.xRot = Mth.sin(f1) * 2.0F;
 			this.Marm.getChild("MarmJoint").getChild("MArm2").xRot = -Mth.sin(f1) * 3.0F;
 		}
